backend/fastapi-backend/pyproject.toml
[project]
name = "fastapi-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "beanie>=1.30.0",
    "dotenv>=0.9.9",
    "fastapi>=0.115.14",
    "motor>=3.7.1",
    "pydantic[email]>=2.11.7",
    "pydantic-settings>=2.10.1",
    "uvicorn>=0.35.0",
    "pdfplumber>=0.11.7",
    "pytesseract>=0.3.13",
    "python-multipart>=0.0.20",
    "python-docx>=1.2.0",
    "numpy>=1.24,<2.0",
    "opencv-python>=4.11.0.86",
    "pdf2image>=1.17.0",
    "sentence-transformers>=4.1.0",
    "scikit-learn>=1.7.0",
    "groq>=0.29.0",
    "pymupdf>=1.26.1",
    "pypdf2>=3.0.1",
    "python-jose[cryptography]>=3.5.0",
    "openai>=1.95.0",
    "spacy>=3.7.2",
    "en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl",
]


[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.ruff]
select = ["E", "F", "I"]
line-length = 88

[dependency-groups]
dev = [
    "black>=25.12.0",
    "isort>=7.0.0",
    "ruff>=0.14.10",
]
